Move declaration of pragma vertex and pixel shader  (#2689)
* Rename LIGHTLOOP_SINGLE_PASS => LIGHTLOOP_DISABLE_TILE_AND_CLUSTER and clean all usage of LIGHTLOOP_TILE_PASS

Rename LIGHTLOOP_SINGLE_PASS to LIGHTLOOP_DISABLE_TILE_AND_CLUSTER
Change multicompile LIGHTLOOP_SINGLE_PASS LIGHTLOOP_TILE_PASS to multicompile _ LIGHTLOOP_DISABLE_TILE_AND_CLUSTER
New behavior, we don't need to define LIGHTLOOP_TILE_PASS anymore to be "tile/cluster". We are tile/cluster by default and need to define LIGHTLOOP_DISABLE_TILE_AND_CLUSTER to have brute force mode.

Disabling tile/cluster is only useful for debugging. Previously the code was commented to allow to debug in LightLoopSettings, This PR re-enable the UI to control it but only allow it for deferred, not forward. This allow to clean the various comment (and it doesn't work with shader graph anyway). We keep debug mode only in the case of deferred.

So in the future we need to debug with brute force mode the lighting, we need to be in deferred and will not able to debug forward material. which should be fine.

* Make Enable Tile And Cluster a deferred only option

* Move pragma vertex and pixel shader to prepare for raytrace shader